+++ CHANGELOG.md
@@ -6,11 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Added Changelog
 - Upgrade to Webpack 3.4.x
 - Upgrade to React 16
 - Enabled HMR
-- Added Preset Manager
-- Update button styles + Added icons
+- Added preset manager
+- Added more presets
+- Update button styles + added icons
 - Update player section UI, including visualiser
 
 ## [1.0.0] - 2017-06-20